Generate sample code from test suite
For each test in the suite, generate a single Markdown file that
contains the highlighted code for each supported language.  This makes
it easy to include those samples in our documentation.

As long as this is the only way to include samples, we can also be
certain that our sample code always works.