Show syntax highlight on index page
When someone first views this site, there's a decent chance that he'll
fiddle with the Language dropdown to see what it does.  If we don't have
any code on this page, the dropdown appears to do nothing.  By putting a
single line code comment here, the dropdown changes the page and the
visitor learns what the dropdown is for.