Make code rendering more consistent
The Go sample for hello-html includes the text "{{.}}" which matches the
escape sequence for Liquid templates.  By wrapping all code samples with
the "{% raw %}" block, accidental Liquid escapes are passed through.

The submodule update in this commit corrects a missing newline character
in one of the Prolog examples.  The absent newline caused some Markdown
to render incorrectly.